refactor: move stuff related to Projects to features/Project folder
