+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "typescript.preferences.importModuleSpecifier": "relative",
+    "typescript.preferences.importModuleSpecifierEnding": "minimal"
+}
