+++ src/components/Assets/AssetsList.tsx
@@ -1,7 +1,10 @@
 import { useDragAndDrop } from "@formkit/drag-and-drop/react";
 import { type FC, useEffect } from "react";
+import type {
+    Asset,
+    AssetKind,
+} from "../../features/Projects/stores/slices/assets";
 import { useAssets } from "../../hooks/useAssets";
-import type { Asset, AssetKind } from "../../stores/storage/assets";
 import type { ResourceProps } from "./AssetsItem";
 import AssetsItem from "./AssetsItem";
