+++ src/components/Playground/Playground.tsx
@@ -3,9 +3,9 @@ import { useEffect, useState } from "react";
 import { useMediaQuery } from "react-responsive";
 import { Slide, ToastContainer } from "react-toastify";
 import { Tooltip } from "react-tooltip";
+import { useProject } from "../../features/Projects/stores/useProject";
 import { useConfig } from "../../hooks/useConfig";
 import { useEditor } from "../../hooks/useEditor.ts";
-import { useProject } from "../../hooks/useProject";
 import { decompressCode } from "../../util/compressCode";
 import { debug } from "../../util/logs";
 import { getPackageInfo } from "../../util/npm.ts";
