+++ src/components/ProjectBrowser/ProjectBrowser.tsx
@@ -7,8 +7,8 @@ import { assets } from "@kaplayjs/crew";
 import { useCallback, useState } from "react";
 import { Tooltip } from "react-tooltip";
 import examplesJson from "../../../kaplay/examples/examples.json";
-import { useProject } from "../../hooks/useProject";
-import type { ProjectMode } from "../../stores/project";
+import type { ProjectMode } from "../../features/Projects/stores/slices/project";
+import { useProject } from "../../features/Projects/stores/useProject";
 import { cn } from "../../util/cn";
 import { TabsList } from "../UI/TabsList";
 import { TabTrigger } from "../UI/TabTrigger";
