+++ src/components/Toolbar/ToolbarToolsMenu.tsx
@@ -3,8 +3,8 @@ import type { FC, PropsWithChildren } from "react";
 import { toast } from "react-toastify";
 import AboutButton from "../../components/About/AboutButton";
 import ConfigOpenDialog from "../../components/Config/ConfigOpenDialog";
+import { useProject } from "../../features/Projects/stores/useProject";
 import { useEditor } from "../../hooks/useEditor";
-import { useProject } from "../../hooks/useProject";
 import { compressCode } from "../../util/compressCode";
 import Projects from "./Projects";
 import ToolbarButton from "./ToolbarButton";
